add to cart with size dish

+++ database/migrations/2024_09_01_113011_create_sizes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('dish_id')->constrained()->onDelete('cascade');
             $table->string('name');
-            $table->decimal('price');
+            $table->float('price');
             $table->timestamps();
         });
     }
